export remade some functions defended

+++ include/libincludes.h
@@ -14,7 +14,7 @@
 # include "libstruct.h"
 
 int ft_execute(t_exec *exec);
-int builtin_call(t_exec *exec);
+ssize_t builtin_call(t_exec *exec);
 int ft_pipe(t_exec *pipe_list);
 int ft_redir_execute(t_exec *exec);
 int ft_no_pipe(t_exec *exec);
